// packages/core/types/data.ts
export type SessionMeta = BaseSessionMeta & { session_id: string };

export interface RolldownData {
  meta: SessionMeta;
  modules: Array<{
    id: string;
    imports?: ModuleImport[];
    importers?: string[];
    build_metrics?: ModuleBuildMetrics;
  }>;
  build_duration: number;
  assets: RolldownAssetInfo[];
  chunks: RolldownChunkInfo[];
  packages: PackageInfo[];
  plugin_build_metrics?: Record<string, PluginBuildMetrics>;
}

export interface BuildMetrics {
  plugin_name: string;
  plugin_id: number;
  duration: number;
  source_code_size?: number;
  transformed_code_size?: number;
  timestamp_start: number;
  timestamp_end: number;
}

export interface ModuleBuildMetrics {
  resolve_ids: RolldownResolveInfo[];
  loads: RolldownModuleLoadInfo[];
  transforms: Array<
    Omit<RolldownModuleTransformInfo, 'diff_added' | 'diff_removed'> & {
      source_code_size: number;
      transformed_code_size: number;
    }
  >;
}

export interface PluginBuildInfo {
  type: 'resolve' | 'load' | 'transform';
  id: string;
  duration: number;
  plugin_id: number;
  plugin_name: string;
  module: string;
  timestamp_start: number;
  timestamp_end: number;
}

export interface PluginBuildMetrics {
  plugin_name: string;
  plugin_id: number;
  calls: (PluginBuildInfo & { unchanged?: boolean })[];
}
// ...
export interface PackageInfo {
  name: string;
  version: string;
  dir: string;
  type?: string;
  transformedCodeSize: number;
  duplicated?: boolean;
  files: {
    path: string;
    transformedCodeSize: number;
    importers: Array<{ path: string; version: string }>;
  }[];
}

export type RolldownPluginBuildMetrics = PluginBuildMetrics & {
  resolveIdMetrics: PluginBuildInfo[];
  loadMetrics: PluginBuildInfo[];
  transformMetrics: PluginBuildInfo[];
};

export interface RolldownResolveInfo {
  type: 'resolve';
  id: string;
  plugin_name: string;
  plugin_id: number;
  importer: string | null;
  module_request: string;
  import_kind: HookResolveIdCallStart['import_kind'];
  resolved_id: string | null;
  timestamp_start: number;
  timestamp_end: number;
  duration: number;
}

export interface RolldownModuleLoadInfo {
  type: 'load';
  id: string;
  plugin_name: string;
  plugin_id: number;
  content: string | null;
  timestamp_start: number;
  timestamp_end: number;
  duration: number;
}

export interface RolldownModuleTransformInfo {
  type: 'transform';
  id: string;
  plugin_name: string;
  plugin_id: number;
  content_from: string | null;
  content_to: string | null;
  diff_added: number;
  diff_removed: number;
  timestamp_start: number;
  timestamp_end: number;
  duration: number;
}

export interface RolldownModuleNoChanges {
  type: 'no_changes_collapsed';
  id: string;
  count: number;
  duration: number;
}

export interface RolldownModuleNoChangesHide {
  type: 'no_changes_hide';
  id: string;
  count: number;
  duration: number;
}

export interface RolldownChunkInfo extends ChunkInfo {
  type?: 'chunk';
  is_initial?: boolean;
  asset?: AssetInfo;
}

export interface RolldownAssetInfo extends AssetInfo {
  type?: 'asset';
  chunk?: RolldownChunkInfo;
}